Fix rounding error in get_mp5_from_spirit.

+++ Character/Class/Druid/Druid.cpp
@@ -73,7 +73,7 @@ unsigned Druid::get_spirit_modifier() const {
 }
 
 unsigned Druid::get_mp5_from_spirit() const {
-    return 15 + static_cast<unsigned>(round(cstats->get_spirit() / 5));
+    return 15 + static_cast<unsigned>(round(static_cast<double>(cstats->get_spirit()) / 5));
 }
 
 double Druid::get_agi_needed_for_one_percent_phys_crit() const {
